removed unnecessary file
